applied additional refactoring to user/community chips
